--- crackle.config.ts
@@ -1,7 +1,0 @@
-import { defineConfig } from '@crackle/cli/config';
-
-export default defineConfig({
-  package: {
-    mode: 'bundle',
-  }
-});
