+++ package.json
@@ -18,50 +18,34 @@
   },
   "license": "MIT",
   "author": "Pete Johns",
-  "exports": {
-    ".": {
-      "types": {
-        "import": "./dist/index.d.mts",
-        "require": "./dist/index.d.ts"
-      },
-      "import": "./dist/index.mjs",
-      "require": "./dist/index.cjs"
-    },
-    "./package.json": "./package.json"
-  },
-  "main": "./dist/index.cjs",
-  "module": "./dist/index.mjs",
-  "types": "./dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
   "scripts": {
-    "build": "pnpm clean ; pnpm lint ; pnpm crackle:package ; pnpm rename && pnpm web-ext:build && pnpm repack-for-chromium",
-    "crackle:package": "crackle package",
-    "clean": "rm -r ./chromium/ ./dist/ ./eventuate-chromium.zip ./web-ext-artifacts/",
-    "rename": "bash -c 'mv ./dist/index.cjs ./dist/index.js'",
+    "build": "webpack",
+    "package": "pnpm lint && pnpm build && pnpm web-ext:lint && pnpm web-ext:build && pnpm repack-for-chromium",
     "repack-for-chromium": "bash './script/repack-for-chromium'",
+    "start": "web-ext run -v -u https://www.parkrun.com.au/brimbank/results/latestresults/",
     "test": "jest",
     "lint": "eslint .",
     "web-ext:build": "web-ext build -o",
-    "web-ext:lint": "web-ext lint",
-    "start": "web-ext run -v -u https://www.parkrun.com.au/coburg/results/latestresults/"
+    "web-ext:lint": "web-ext lint"
   },
   "devDependencies": {
-    "@crackle/cli": "^0.15.5",
     "@eslint/js": "^9.13.0",
     "@types/eslint__js": "^8.42.3",
     "@types/jest": "^29.5.13",
-    "@types/node": "^22.7.5",
+    "@types/node": "^22.7.7",
+    "copy-webpack-plugin": "^12.0.2",
     "eslint": "^9.13.0",
     "globals": "^15.11.0",
     "jest": "^29.7.0",
     "prettier": "^3.3.3",
     "ts-jest": "^29.2.5",
+    "ts-loader": "^9.5.1",
     "ts-node-dev": "^2.0.0",
-    "typescript": "^5.6.2",
+    "typescript": "^5.6.3",
     "typescript-eslint": "^8.10.0",
-    "web-ext": "^8.3.0"
+    "web-ext": "^8.3.0",
+    "webpack": "^5.95.0",
+    "webpack-cli": "^5.1.4"
   },
   "packageManager": "pnpm@9.2.0",
   "webExt": {
@@ -73,9 +57,6 @@
       "LICENSE.txt",
       "README.md",
       "chromium*",
-      "crackle.config.ts",
-      "dist/index.mjs",
-      "dist/index.d.*",
       "jest.config.ts",
       "package.json",
       "pnpm-lock.yaml",
