+++ src/index.ts
@@ -2,9 +2,11 @@ import { ResultsPageExtractor } from "./extractors/ResultsPageExtractor";
 import { MilestonePresenter } from "./presenters/MilestonePresenter";
 import { conjoin, pluralize, sortAndConjoin } from "./stringFunctions";
 import { fiveKFinishersToMilestones } from "./transformers/fiveKFinishersToMilestones";
-import { Volunteer } from "./types/Volunteer";
 import { upsertParagraph } from "./dom/upsertParagraph";
+import { presentVolunteerName } from "./presenters/presentVolunteerName";
+import { sourceVolunteerCount } from "./extractors/sourceVolunteerCount";
 
+function eventuate(): void {
 const rpe = new ResultsPageExtractor(document);
 
 const milestoneCelebrations = fiveKFinishersToMilestones(rpe.finishers);
@@ -133,72 +135,6 @@ if (insertionPoint) {
       `We shall always be grateful to each of our ${rpe.facts.volunteers.toLocaleString()} wonderful volunteers for their contributions.`
   );
 }
-
-function sourceVolunteerCount(v: Volunteer, update: HTMLSpanElement) {
-  const timeout = v.athleteID % 1000;
-  const volunteerUrl = new URL(
-    `/parkrunner/${v.athleteID}/`,
-    window.location.origin
-  ).toString();
-
-  setTimeout(() => {
-    fetch(volunteerUrl)
-      .then((r) => r.text())
-      .then((html) => new DOMParser().parseFromString(html, "text/html"))
-      .then((doc) => {
-        return {
-          vols: doc.querySelector(
-            "h3#volunteer-summary + table tfoot td:last-child"
-          ),
-          agegroup: doc.querySelector("h3 + p") as HTMLParagraphElement,
-        };
-      })
-      .then((e) => {
-        v.vols = e.vols?.textContent ?? "";
-        v.agegroup =
-          e.agegroup?.textContent?.trim().split(" ").slice(-1)[0] ?? "";
-
-        update.innerText = presentVolunteerName(
-          v.name,
-          Number(v.vols),
-          v.agegroup
-        );
-        update.dataset.vols = v.vols;
-        update.dataset.agegroup = v.agegroup;
-        update.dataset.vols_source = volunteerUrl;
-        const a: HTMLAnchorElement | null = document.querySelector(
-          `a[data-athleteid="${v.athleteID}"]`
-        );
-        if (a) {
-          a.dataset.vols = v.vols;
-          a.dataset.agegroup = v.agegroup;
-          a.dataset.vols_source = volunteerUrl;
-        }
-      });
-  }, timeout);
 }
 
-function presentVolunteerName(
-  name: string,
-  vols: number,
-  agegroup: string
-): string {
-  const milestones: Record<number, string> = {
-    10: "J",
-    25: "",
-    50: "",
-    100: "",
-    250: "",
-    500: "",
-    1000: "",
-  };
-
-  for (const n in milestones) {
-    const restricted_age: string = milestones[Number(n)];
-    if (vols === Number(n) && agegroup.startsWith(restricted_age)) {
-      return `${name} (congratulations on joining the v${n}-club)`;
-    }
-  }
-
-  return name;
-}
+eventuate();
