+++ tsconfig.json
@@ -1,13 +1,13 @@
 {
   "compilerOptions": {
     "target": "ESNext",
-    "module": "System",
+    "module": "CommonJS",
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
     "strict": true,
     "skipLibCheck": true,
-    "rootDir": "src",
-    "outFile": "./dist/eventuate.js"
+    "rootDir": "./src",
+    "outDir": "./dist"
   },
   "exclude": ["*.config.ts", "test"]
 }
