+++ webpack.config.js
@@ -1,0 +1,25 @@
+const path = require('path');
+
+module.exports = {
+  mode: 'production',
+  entry: './src/index.ts',
+  module: {
+    rules: [
+      {
+        test: /\.tsx?$/,
+        use: 'ts-loader',
+        exclude: /node_modules/,
+      },
+    ],
+  },
+  resolve: {
+    extensions: ['.ts', '.js'],
+  },
+  output: {
+    filename: 'eventuate.js',
+    path: path.resolve(__dirname, 'dist'),
+  },
+  optimization: {
+    minimize: false
+  }
+};
